Refactor
- Refactor: Move the Stack-In-A-Box exceptions to their own module
  these are all internal so should not affect users
- Refactor: Move the Stack-In-A-Box router to its own module,
  this is internal so should not affect users
- Bug Fix: add the `about` page to the toctree for the docs
- Bug Fix: pycodestyle fixes
