+++ docs/api/services/service-exceptions.rst
@@ -6,7 +6,7 @@ StackInABoxService Exceptions
 StackInABoxService has several custom exceptions which may come up during
 service and route registration.
 
-.. currentmodule:: stackinabox.services.service
+.. currentmodule:: stackinabox.services.exceptions
 .. autoexception:: StackInABoxServiceErrors
 .. autoexception:: RouteAlreadyRegisteredError
 .. autoexception:: InvalidRouteRegexError
