+++ docs/community/index.rst
@@ -7,3 +7,4 @@ Community Guide
     help
     contribute
     faq
+    about
